+++ src/pendoEnv.ts
@@ -1,6 +1,9 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
+import { filterGuideTag, filterResponse, isTransmissionAllowed } from './pendoFilters'
+import { createTransmissions } from './pendoTransmissions'
+
 type UnderscoreLike = {
   each: (...args: any) => unknown
   keys: (...args: any) => (string | symbol)[]
@@ -12,26 +15,6 @@ function makeError(msg: string): Error {
   msg = `PendoEnv: ${msg}`
   console.error(msg)
   return new Error(msg)
-}
-
-function filterGuideTag(tagName: string, attributes: Record<string, string>) {
-  tagName = tagName.toLowerCase().trim()
-  attributes = Object.fromEntries(
-    Object.entries(attributes).map(([k, v]) => [k.toLowerCase().trim(), v])
-  )
-  switch (tagName) {
-    case 'embed':
-    case 'iframe':
-    case 'script':
-      throw makeError(`guides may not contain '${tagName}'`)
-    case 'a':
-      if (attributes.href && /^\s*javascript:/i.test(attributes.href)) {
-        throw makeError(`guides may not contain 'javascript:' links`)
-      }
-      break
-    default:
-      return
-  }
 }
 
 function getTransmissionData(
@@ -60,151 +43,6 @@ function getTransmissionData(
   }
 }
 
-function isTransmissionAllowed(
-  url: URL,
-  data: object | undefined,
-  integrity: string | undefined,
-  params: {
-    apiKey: string
-    VERSION: string
-    transmissionLog: Array<{ endpoint: string } & Record<string, unknown>>
-    pendoOptions: Record<string, unknown>
-  }
-) {
-  // This is excessively paranoid, but it limits the data leakage possible to a 1-byte
-  // range of values. (Actual values are 0-1.)
-  const ctEpsilon = 128
-  const dataHost = params.pendoOptions.dataHost
-  const guideHosts = ((params.pendoOptions.allowedOriginServers as string[]) ?? []).map(
-    (x) => new URL(x).host
-  )
-
-  if (url.protocol !== 'https:') {
-    throw makeError(`fetch to non-https url not allowed`)
-  }
-  // This is all a bit overly restrictive, but it should ensure we fail fast if
-  // any of the assumptions made are incorrect.
-  if (dataHost && url.host === dataHost) {
-    const [match, endpoint, apiKey] = /^\/data\/([^/]*)\/(.*)$/.exec(url.pathname) ?? []
-    if (!match) throw makeError(`fetch from data.pendo.io does not match expected regex`)
-    if (apiKey !== params.apiKey) {
-      throw makeError(`expected api key in url to match config (${apiKey})`)
-    }
-    // Verify no unexpected data in the URL parameters
-    for (const [k, v] of url.searchParams.entries()) {
-      switch (k) {
-        case 'jzb':
-          break
-        case 'v':
-          if (v !== params.VERSION) {
-            throw makeError(
-              `PendoEnv: attempted fetch with url parameter 'v' which does not match agent version`
-            )
-          }
-          break
-        case 'ct': {
-          const ct = Number.parseInt(v)
-          if (
-            !Number.isSafeInteger(ct) ||
-            ct.toString() !== v ||
-            Math.abs(ct - Date.now()) > ctEpsilon
-          ) {
-            throw makeError(
-              `PendoEnv: attempted fetch with url parameter 'ct' out of expected range: ${v}`
-            )
-          }
-          console.debug(`PendoEnv: ct diff`, Math.abs(ct - Date.now()))
-          break
-        }
-        default:
-          throw makeError(`attempted fetch with unexpected url parameter '${k}' = '${v}'`)
-      }
-    }
-    // Log the transmission
-    params.transmissionLog.push(
-      ...(Array.isArray(data) ? data : [data]).map((x) => ({
-        endpoint,
-        ...x
-      }))
-    )
-  } else if (guideHosts.includes(url.host)) {
-    if (!/^\/guide-content\/.*\.dom\.json$/.test(url.pathname)) {
-      throw makeError(
-        `fetch from pendo-static-6047664892149760.storage.googleapis.com does not match expected regex`
-      )
-    }
-    // Verify no unexpected data in the URL parameters
-    let sawIntegrity = false
-    for (const [k, v] of url.searchParams.entries()) {
-      switch (k) {
-        case 'sha256': {
-          if (`sha256-${v}` !== integrity) {
-            throw makeError(`expected integrity url parameter to match request's integrity value`)
-          }
-          sawIntegrity = true
-          break
-        }
-        default:
-          throw makeError(`attempted fetch with unexpected url parameter '${k}' = '${v}'`)
-      }
-    }
-    if (!sawIntegrity) {
-      throw makeError(`expected integrity url parameter on request`)
-    }
-  } else {
-    throw makeError(`agent tried to fetch an unrecognized url (${url})`)
-  }
-}
-
-async function filterResponse(
-  url: URL,
-  data: object | undefined,
-  response: Response
-): Promise<Response> {
-  if (response.status < 200 || response.status >= 300) return Response.error()
-  if (!response.body) return new Response(null, { status: 200 })
-  const resObj = await response.json()
-  if (Array.isArray(resObj)) {
-    makeError(`fetch response is an array, not an object: ${JSON.stringify(resObj, undefined, 2)}`)
-    return Response.error()
-  }
-  const unexpectedKeys = Object.keys(resObj).filter(
-    (x) =>
-      ![
-        'autoOrdering',
-        'designerEnabled',
-        'features',
-        'globalJsUrl',
-        'guideCssUrl',
-        'guideWidget',
-        'guides',
-        'lastGuideStepSeen',
-        'normalizedUrl',
-        'preventCodeInjection',
-        'segmentFlags',
-        'throttling'
-      ].includes(x)
-  )
-  if (unexpectedKeys.length > 0) {
-    makeError(`fetch response has unexpected keys: ${JSON.stringify(unexpectedKeys, undefined, 2)}`)
-    return Response.error()
-  }
-  // Override preventCodeInjection to true in case the server feels like trying
-  // to unset it
-  if ('guides' in resObj || 'preventCodeInjection' in resObj) {
-    if (resObj.preventCodeInjection !== true) {
-      console.warn(
-        `PendoEnv: Expected preventCodeInjection to be set on a guide, but it wasn't. It's been set anyway.`
-      )
-    }
-    resObj.preventCodeInjection = true
-  }
-  // The agent doesn't use headers or statusText, so we can ignore them
-  return new Response(JSON.stringify(resObj), {
-    status: response.status
-  })
-}
-
 export function makePendoEnv(pendoOptions: Record<string, unknown>) {
   // While the state object has a variety of values that start out as undefined,
   // they are usually set during parsing of the agent script and guaranteed to
@@ -227,7 +65,7 @@ export function makePendoEnv(pendoOptions: Record<string, unknown>) {
     VERSION: undefined as string | undefined
   }
 
-  const transmissionLog: Array<{ endpoint: string } & Record<string, unknown>> = []
+  const transmissions = createTransmissions()
 
   function fetchAndLog(url: string, init?: RequestInit): Promise<Response> {
     const urlObj = new URL(url)
@@ -243,8 +81,8 @@ export function makePendoEnv(pendoOptions: Record<string, unknown>) {
     isTransmissionAllowed(urlObj, dataObj, init?.integrity, {
       apiKey: state.apiKey!,
       VERSION: state.VERSION!,
-      transmissionLog: transmissionLog,
-      pendoOptions
+      pendoOptions,
+      transmissions
     })
 
     return fetch(url, init)
@@ -384,6 +222,9 @@ export function makePendoEnv(pendoOptions: Record<string, unknown>) {
       if (state.VERSION && state.VERSION !== value) {
         throw makeError(`only expected VERSION to be set once`)
       }
+      if (!state.VERSION && value !== '2.117.0_prod') {
+        console.warn(`PendoEnv: Unexpected agent version; it may break in very unexpected ways.`)
+      }
       state.VERSION = value
     },
     get apiKey() {
@@ -404,6 +245,7 @@ export function makePendoEnv(pendoOptions: Record<string, unknown>) {
       _headers = new Headers()
       _url = ''
       _method = ''
+      _sha256 = null as string | null
       withCredentials = false
       status = 0
       readyState = 0
@@ -412,6 +254,7 @@ export function makePendoEnv(pendoOptions: Record<string, unknown>) {
       open(method: string, url: string) {
         this._method = method
         this._url = url
+        this._sha256 = new URL(url).searchParams.get('sha256')
       }
       setRequestHeader(name: string, value: string) {
         this._headers.append(name, value)
@@ -422,7 +265,8 @@ export function makePendoEnv(pendoOptions: Record<string, unknown>) {
           method: this._method,
           headers: this._headers,
           credentials: this.withCredentials ? 'include' : 'same-origin',
-          body: data
+          body: data,
+          integrity: this._sha256 ? `sha256-${this._sha256}` : undefined
         })
           .then(async (res: Response) => {
             this.status = res.status
@@ -484,7 +328,7 @@ export function makePendoEnv(pendoOptions: Record<string, unknown>) {
     pendo,
     // The agent doesn't use this, but it makes the log relatively easy to find
     // as window.pendoEnv.transmissionLog.
-    transmissionLog,
+    transmissions,
     window: new Proxy(window, {
       get(target, p) {
         // We undefine window.fetch and window.XMLHttpRequest to force the agent
